requirements.txt
aiohttp>=3.8.1
aiosignal>=1.2.0
async-timeout>=4.0.2
attrs>=21.4.0
charset-normalizer>=2.0.10
colorama>=0.4.4
colorlog>=6.6.0
frozenlist>=1.3.0
halo>=0.0.31
hikari>=2.0.0.dev105
hikari-lightbulb>=2.1.3
idna>=3.3
log-symbols>=0.0.14
multidict>=5.2.0
pure25519>=0.0.1
python-dotenv>=0.19.2
python-weather>=0.3.7
six>=1.16.0
spinners>=0.0.24
termcolor>=1.1.0
xmltodict>=0.12.0
yarl>=1.7.2
uvloop>=0.16; sys_platform != "win32"